requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
ruamel.yaml==0.18.6
python-multipart==0.0.9